Add logic for step 3 requirements
What was changed?
-----------------------
This commit adds all the logic to implement the DocumentService, QdrantService, and FastAPI endpoints.
It also adds a `run.py` script to launch the service.
Some logic may be refactored before final submission.

Why was this change made?
-----------------------
These changes were made to meet the functional requirements for the Norm AI take-home project.
The `run.py` script was added due to an unsolved bug causing the `app` module not to be recognized in the build.
Suspected cause is an issue with the PYTHONPATH env variable.

Testing
-----------------------
- Ran the service locally and sent queries via curl
- Verified correct parsing of docs via logging and query output